+++ example/android/app/src/main/kotlin/vn/vietmap/demo_plugin_example/MainActivity.kt
@@ -1,7 +1,6 @@
 package vn.vietmap.demo_plugin_example
 
 import io.flutter.embedding.android.FlutterActivity
-import io.flutter.plugin.common.PluginRegistry.Registrar
 
 class MainActivity: FlutterActivity() {